CIRCLE + FOLLOWING !!!! READ
Implemented a lot in this commit:

Now the grid can be efficiently spawned with a round shape.
The cubes can now follow the player for infinite procedural generation of terrain!
Script allows user to choose how the curve behaves via inspector

To do:
Delete the cubes trace when the player moves (optional to check via inspector) in order to ALWAYS have a circle that moves below, deleteng old and creating new blocks